fix: add missing JobSpec fields and fix MaxWorkers type
- Add RepoURL, RepoBranch, ProjectRoot, RunType to runner.JobSpec
  (worker was already setting them, struct was just missing the fields)
- Pass the new fields through to container env vars
- Remove int32 cast on RunnerMaxConcurrent — river.QueueConfig.MaxWorkers
  is int, not int32

+++ api/internal/worker/worker.go
@@ -47,7 +47,7 @@ func New(pool *pgxpool.Pool, cfg *config.Config, r *runner.Runner, s *storage.Cl
 
 	rc, err := river.NewClient(riverpgxv5.New(pool), &river.Config{
 		Queues: map[string]river.QueueConfig{
-			river.QueueDefault: {MaxWorkers: int32(cfg.RunnerMaxConcurrent)},
+			river.QueueDefault: {MaxWorkers: cfg.RunnerMaxConcurrent},
 		},
 		Workers: workers,
 	})
